pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "biometric-mlops"
version = "0.1.0"
description = "Simple multimodal biometric training project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "torch",
  "pandas",
  "numpy",
  "pyyaml",
  "mlflow",
  "Pillow",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["src", "src.*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
